SCHED-187: Circular dependency added
