+++ app/process_manager/manager.py
@@ -2,7 +2,7 @@ import asyncio
 import signal
 from random import randint
 from multiprocessing import Process
-from common.meta import Singleton
+from app.common.meta import Singleton
 from datetime import datetime
 from .runner import StandardRunner
 from .task import SchedulerTask, TaskType
