+++ main.py
@@ -1,11 +1,9 @@
 import uvicorn
 from fastapi.responses import JSONResponse
-from dotenv import load_dotenv
-import os
-from omegaconf import OmegaConf
 from app.graphql.schema import PlanManager
 from app.process_manager import ProcessManager, TaskType
 from app.scheduler import Scheduler
+from app.config import config
 from app import app
 from astropy.time import Time
 from datetime import datetime
@@ -18,8 +16,6 @@ async def new_schedule(manager, scheduler):
     manager.add_task(datetime.now(), scheduler, TaskType.STANDARD)
     await asyncio.sleep(10)
 
-# session = Session(config.graphql.url)
-
 # Root API
 @app.get("/", include_in_schema=False)
 def root() -> JSONResponse:
@@ -28,7 +24,7 @@ def root() -> JSONResponse:
                         content={
                             "message": "Welcome to Server"})
 
-
+'''
 @app.get("/new", include_in_schema=False)
 def new() -> JSONResponse:
     start = Time("2018-10-01 08:00:00", format='iso', scale='utc')
@@ -37,21 +33,10 @@ def new() -> JSONResponse:
     asyncio.run(new_schedule(manager, scheduler))
     return JSONResponse(status_code=200,
                         content={ "message": f"New schedule created: {plan_manager.get_plans()}"})
+'''
+
 
 if __name__ == "__main__":
-    load_dotenv()
-
-    path = os.path.join(os.getcwd(), '.', 'config.yaml')
-
-    config = OmegaConf.load(path)
-    config.graphql.url = os.environ.get("GRAPHQL_URL")  # HACK: to get the url from the env
-
     manager = ProcessManager(size=config.process_manager.size,
                              timeout=config.process_manager.timeout)
-    
-
-    # Run random schedule by process manager
-    #  mode = TaskType.STANDARD
-
-    #print(plan_manager.get_plans())
     uvicorn.run(app, host='127.0.0.1', port=8000)
